refactor test_activation_defence_pytest for pytest
